Using ArticleList as StartActivity for normal app launches. - adding Buttons to sync and for Settings to ActionBar
